+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This file is used to list changes made in each version of cookbook-openstack-com
 * Removed all ceph logic
 * Allow different URLs for internal, public, admin endpoints (bug 1412919)
 * Cleanup CHEF-3694 warnings
+* Allow disabling GnuPG signature check for RDO repo
 
 ## 10.1.0
 * Adding identity admin bind host endpoint to allow flexibility and consistency